+++ modules/cpdos/hmc.py
@@ -6,7 +6,7 @@ Attempts to find Cache Poisoning with HTTP Metachar Character (HMC)
 https://cpdos.org/#HMC
 """
 
-from modules.utils import random, requests, configure_logger, human_time
+from modules.utils import random, requests, configure_logger, human_time, Identify
 
 logger = configure_logger(__name__)
 
@@ -40,7 +40,7 @@ def check_meta_character(url, s, main_status_code, authent, meta_character, huma
     if reason:
         payload = f"PAYLOAD: {headers}"
         print(
-            f"\033[31m └── [VULNERABILITY CONFIRMED]\033[0m | HMC | \033[34m{url}\033[0m | {reason} | {payload}"
+            f" {Identify.confirmed} | HMC | \033[34m{url}\033[0m | {reason} | {payload}"
         )
     human_time(human)